+++ Scripts/Items/Champion Artifacts/Shared/GuantletsOfAnger.cs
@@ -7,8 +7,8 @@ namespace Server.Items
 	{
 		public override int LabelNumber{ get{ return 1094902; } } // Gauntlets of Anger [Replica]
 
-		public override int BasePhysicalResistance{ get{ return 4; } }
-		public override int BaseMagieResistance{ get{ return 5; } }
+		public override double BasePhysicalResistance{ get{ return 4; } }
+		public override double BaseMagieResistance{ get{ return 5; } }
 
 		public override int InitMinHits{ get{ return 150; } }
 		public override int InitMaxHits{ get{ return 150; } }
