+++ Scripts/Items/Minor Artifacts/GlovesOfThePugilist.cs
@@ -7,7 +7,7 @@ namespace Server.Items
 	{
 		public override int LabelNumber{ get{ return 1070690; } }
 
-		public override int BasePhysicalResistance{ get{ return 18; } }
+		public override double BasePhysicalResistance{ get{ return 18; } }
 
 		public override int InitMinHits{ get{ return 255; } }
 		public override int InitMaxHits{ get{ return 255; } }
