+++ Scripts/Items/Minor Artifacts/VioletCourage.cs
@@ -7,8 +7,8 @@ namespace Server.Items
 	{
 		public override int LabelNumber{ get{ return 1063471; } }
 
-		public override int BasePhysicalResistance{ get{ return 14; } }
-		public override int BaseMagieResistance{ get{ return 9; } }
+		public override double BasePhysicalResistance{ get{ return 14; } }
+		public override double BaseMagieResistance{ get{ return 9; } }
 
 		public override int InitMinHits{ get{ return 255; } }
 		public override int InitMaxHits{ get{ return 255; } }
