+++ Scripts/Items/Temrael/Armures/ArmureElfe.cs
@@ -6,8 +6,8 @@ namespace Server.Items
     public class ElfeHelm : BaseArmor
     {
 
-        public override int BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorFeuilles.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorFeuilles.max_Durabilite; } }
@@ -46,8 +46,8 @@ namespace Server.Items
     public class ElfeGorget : BaseArmor
     {
 
-        public override int BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorFeuilles.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorFeuilles.max_Durabilite; } }
@@ -86,8 +86,8 @@ namespace Server.Items
     public class ElfeArms : BaseArmor
     {
 
-        public override int BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorFeuilles.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorFeuilles.max_Durabilite; } }
@@ -126,8 +126,8 @@ namespace Server.Items
     public class ElfeLeggings : BaseArmor
     {
 
-        public override int BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorFeuilles.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorFeuilles.max_Durabilite; } }
@@ -166,8 +166,8 @@ namespace Server.Items
     public class ElfeTunic : BaseArmor
     {
 
-        public override int BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorFeuilles.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorFeuilles.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorFeuilles.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorFeuilles.max_Durabilite; } }
